make codebase more modular
