+++ fsrs/__init__.py
@@ -5,25 +5,10 @@ py-fsrs
 Py-FSRS is the official Python implementation of the FSRS scheduler algorithm, which can be used to develop spaced repetition systems.
 """
 
-from .fsrs import (
-    Scheduler,
-    Card,
-    Rating,
-    ReviewLog,
-    State,
-    DEFAULT_PARAMETERS,
-    STABILITY_MIN,
-)
+from fsrs.scheduler import Scheduler
+from fsrs.card import Card, State
+from fsrs.review_log import ReviewLog, Rating
 
-from .optimizer import Optimizer
+from fsrs.optimizer import Optimizer
 
-__all__ = [
-    "Scheduler",
-    "Card",
-    "Rating",
-    "ReviewLog",
-    "State",
-    "Optimizer",
-    "DEFAULT_PARAMETERS",
-    "STABILITY_MIN",
-]
+__all__ = ["Scheduler", "Card", "Rating", "ReviewLog", "State", "Optimizer"]
