+++ fsrs/scheduler.py
@@ -1,14 +1,10 @@
 """
-fsrs.fsrs
+fsrs.scheduler
 ---------
 
-This module defines each of the classes used in the fsrs package.
+This module defines the Scheduler class as well as the various constants used in its calculations.
 
 Classes:
-    State: Enum representing the learning state of a Card object.
-    Rating: Enum representing the four possible ratings when reviewing a card.
-    Card: Represents a flashcard in the FSRS system.
-    ReviewLog: Represents the log entry of a Card that has been reviewed.
     Scheduler: The FSRS spaced-repetition scheduler.
 """
 
@@ -17,10 +13,10 @@ from collections.abc import Sequence
 import math
 from datetime import datetime, timezone, timedelta
 from copy import copy
-from enum import IntEnum
 from random import random
-import time
 from dataclasses import dataclass
+from fsrs.card import Card, State
+from fsrs.review_log import ReviewLog, Rating
 
 DEFAULT_PARAMETERS = (
     0.2172,
@@ -116,208 +112,6 @@ FUZZ_RANGES = [
         "factor": 0.05,
     },
 ]
-
-
-class State(IntEnum):
-    """
-    Enum representing the learning state of a Card object.
-    """
-
-    Learning = 1
-    Review = 2
-    Relearning = 3
-
-
-class Rating(IntEnum):
-    """
-    Enum representing the four possible ratings when reviewing a card.
-    """
-
-    Again = 1
-    Hard = 2
-    Good = 3
-    Easy = 4
-
-
-@dataclass(init=False)
-class Card:
-    """
-    Represents a flashcard in the FSRS system.
-
-    Attributes:
-        card_id: The id of the card. Defaults to the epoch milliseconds of when the card was created.
-        state: The card's current learning state.
-        step: The card's current learning or relearning step or None if the card is in the Review state.
-        stability: Core mathematical parameter used for future scheduling.
-        difficulty: Core mathematical parameter used for future scheduling.
-        due: The date and time when the card is due next.
-        last_review: The date and time of the card's last review.
-    """
-
-    card_id: int
-    state: State
-    step: int | None
-    stability: float | None
-    difficulty: float | None
-    due: datetime
-    last_review: datetime | None
-
-    def __init__(
-        self,
-        card_id: int | None = None,
-        state: State = State.Learning,
-        step: int | None = None,
-        stability: float | None = None,
-        difficulty: float | None = None,
-        due: datetime | None = None,
-        last_review: datetime | None = None,
-    ) -> None:
-        if card_id is None:
-            # epoch milliseconds of when the card was created
-            card_id = int(datetime.now(timezone.utc).timestamp() * 1000)
-            # wait 1ms to prevent potential card_id collision on next Card creation
-            time.sleep(0.001)
-        self.card_id = card_id
-
-        self.state = state
-
-        if self.state == State.Learning and step is None:
-            step = 0
-        self.step = step
-
-        self.stability = stability
-        self.difficulty = difficulty
-
-        if due is None:
-            due = datetime.now(timezone.utc)
-        self.due = due
-
-        self.last_review = last_review
-
-    def to_dict(self) -> dict[str, int | float | str | None]:
-        """
-        Returns a JSON-serializable dictionary representation of the Card object.
-
-        This method is specifically useful for storing Card objects in a database.
-
-        Returns:
-            A dictionary representation of the Card object.
-        """
-
-        return_dict = {
-            "card_id": self.card_id,
-            "state": self.state.value,
-            "step": self.step,
-            "stability": self.stability,
-            "difficulty": self.difficulty,
-            "due": self.due.isoformat(),
-            "last_review": self.last_review.isoformat() if self.last_review else None,
-        }
-
-        return return_dict
-
-    @staticmethod
-    def from_dict(source_dict: dict[str, int | float | str | None]) -> Card:
-        """
-        Creates a Card object from an existing dictionary.
-
-        Args:
-            source_dict: A dictionary representing an existing Card object.
-
-        Returns:
-            A Card object created from the provided dictionary.
-        """
-
-        card_id = int(source_dict["card_id"])
-        state = State(int(source_dict["state"]))
-        step = source_dict["step"]
-        stability = (
-            float(source_dict["stability"]) if source_dict["stability"] else None
-        )
-        difficulty = (
-            float(source_dict["difficulty"]) if source_dict["difficulty"] else None
-        )
-        due = datetime.fromisoformat(source_dict["due"])
-        last_review = (
-            datetime.fromisoformat(source_dict["last_review"])
-            if source_dict["last_review"]
-            else None
-        )
-
-        return Card(
-            card_id=card_id,
-            state=state,
-            step=step,
-            stability=stability,
-            difficulty=difficulty,
-            due=due,
-            last_review=last_review,
-        )
-
-
-@dataclass
-class ReviewLog:
-    """
-    Represents the log entry of a Card object that has been reviewed.
-
-    Attributes:
-        card_id: The id of the card being reviewed.
-        rating: The rating given to the card during the review.
-        review_datetime: The date and time of the review.
-        review_duration: The number of miliseconds it took to review the card or None if unspecified.
-    """
-
-    card_id: int
-    rating: Rating
-    review_datetime: datetime
-    review_duration: int | None
-
-    def to_dict(
-        self,
-    ) -> dict[str, dict | int | str | None]:
-        """
-        Returns a JSON-serializable dictionary representation of the ReviewLog object.
-
-        This method is specifically useful for storing ReviewLog objects in a database.
-
-        Returns:
-            A dictionary representation of the ReviewLog object.
-        """
-
-        return_dict = {
-            "card_id": self.card_id,
-            "rating": self.rating.value,
-            "review_datetime": self.review_datetime.isoformat(),
-            "review_duration": self.review_duration,
-        }
-
-        return return_dict
-
-    @staticmethod
-    def from_dict(
-        source_dict: dict[str, dict | int | str | None],
-    ) -> ReviewLog:
-        """
-        Creates a ReviewLog object from an existing dictionary.
-
-        Args:
-            source_dict: A dictionary representing an existing ReviewLog object.
-
-        Returns:
-            A ReviewLog object created from the provided dictionary.
-        """
-
-        card_id = source_dict["card_id"]
-        rating = Rating(int(source_dict["rating"]))
-        review_datetime = datetime.fromisoformat(source_dict["review_datetime"])
-        review_duration = source_dict["review_duration"]
-
-        return ReviewLog(
-            card_id=card_id,
-            rating=rating,
-            review_datetime=review_datetime,
-            review_duration=review_duration,
-        )
 
 
 @dataclass(init=False)
@@ -930,3 +724,6 @@ class Scheduler:
         fuzzed_interval = timedelta(days=fuzzed_interval_days)
 
         return fuzzed_interval
+
+
+__all__ = ["Scheduler"]
